+++ src/objects/solids/Cylinder.h
@@ -10,23 +10,23 @@
 
 class Cylinder : public Solid {
 public:
-    Cylinder(const Point& cb, TexMat& tex, const Vector& d, float r) : Solid(cb, tex), cb_(cb), r_(r) {
+    Cylinder(const Point& cb, TexMat& tex, const Vector& d, double r) : Solid(cb, tex), cb_(cb), r_(r) {
         cd_ = d.normalized();
         h_ = d.magnitude();
         ct_ = cb_ + h_ * cd_;
     }
 
-    float intersects(const Line& line) const override;
+    double intersects(const Line& line) const override;
     Line get_normal(const Point& p) const override;
     TexPixel get_tex(const Point& p) const override;
 private:
     Point cb_;
-    float r_;
+    double r_;
     Vector cd_;
-    float h_;
+    double h_;
     Point ct_;
 
-    float intersects_base(const Line& line, bool bottom) const;
+    double intersects_base(const Line& line, bool bottom) const;
 };
